# openapi.json
{"openapi": "3.0.2", "info": {"title": "falcon-connector", "version": "0.1.0"}, "paths": {"/metrics": {"get": {"summary": "Metrics", "description": "Endpoint that serves Prometheus metrics.", "operationId": "metrics_metrics_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/specification/list": {"get": {"tags": ["connector specification endpoints"], "summary": "Get Connector Spec", "operationId": "get_connector_spec_api_v1_specification_list_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Connector Spec Api V1 Specification List Get"}}}}}}}, "/api/v1/specification/test": {"post": {"tags": ["connector specification endpoints"], "summary": "Check Connector Spec", "operationId": "check_connector_spec_api_v1_specification_test_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Check Connector Spec Api V1 Specification Test Post"}}}}}}}, "/api/v1/detections/detections": {"get": {"tags": ["detections"], "summary": "List Detections", "description": "List all detections. Results can be customized by passing a FQL filter.\nFilters can be provided using the -f argument, allowing you to reduce the number of results returned and filter down to just the records of interest to you.\nWill not work -> device.hostname:*'*search-string*'\nWill work -> \"device.hostname:*'*search-string*'\"", "operationId": "list_detections_api_v1_detections_detections_get", "parameters": [{"required": false, "schema": {"title": "Search Filter", "type": "string"}, "name": "search_filter", "in": "query"}, {"required": false, "schema": {"title": "Limit", "type": "integer", "default": 10}, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response List Detections Api V1 Detections Detections Get", "type": "array", "items": {}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/detections/detections/{detection_id}": {"get": {"tags": ["detections"], "summary": "Get Detection By Id", "description": "Get detection by a single id. Multiple IDs can be specified by delimiting with comma (ID1,ID2,ID3).\nA maximum of 20 IDs may be specified in this manner.", "operationId": "get_detection_by_id_api_v1_detections_detections__detection_id__get", "parameters": [{"required": true, "schema": {"title": "Detection Id", "type": "string"}, "name": "detection_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Detection By Id Api V1 Detections Detections  Detection Id  Get", "type": "array", "items": {}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/detections/get-aggregate-detects": {"get": {"tags": ["detections"], "summary": "Get Aggregate Detects", "description": "Get detect aggregates as specified via json in request body.", "operationId": "get_aggregate_detects_api_v1_detections_get_aggregate_detects_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Aggregate Detects Api V1 Detections Get Aggregate Detects Get", "type": "array", "items": {}}}}}}}}, "/api/v1/detections/get-detect-summary": {"get": {"tags": ["detections"], "summary": "Get Detect Summary", "description": "View information about detections for all ids listed in the id_list", "operationId": "get_detect_summary_api_v1_detections_get_detect_summary_get", "parameters": [{"required": true, "schema": {"title": "Id List", "type": "string"}, "name": "id_list", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Detect Summary Api V1 Detections Get Detect Summary Get", "type": "array", "items": {}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/detections/update-detects-by-ids": {"get": {"tags": ["detections"], "summary": "Update Detects By Ids", "description": "Update the detection to the provided status.", "operationId": "update_detects_by_ids_api_v1_detections_update_detects_by_ids_get", "parameters": [{"required": true, "schema": {"title": "Id List", "type": "string"}, "name": "id_list", "in": "query"}, {"required": true, "schema": {"title": "Status", "type": "string"}, "name": "status", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Update Detects By Ids Api V1 Detections Update Detects By Ids Get", "type": "array", "items": {}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}